fix emin,emax discovery for gtlike flux calculations in binned case; suppress double->int warnings in SourceMap; enable CountsMap to set the image data using a vector<couble>

+++ src/SourceMap.cxx
@@ -4,7 +4,7 @@
  *        response.
  * @author J. Chiang
  *
- * $Header: /nfs/slac/g/glast/ground/cvs/Likelihood/src/SourceMap.cxx,v 1.74 2010/02/17 07:12:16 jchiang Exp $
+ * $Header: /nfs/slac/g/glast/ground/cvs/Likelihood/src/SourceMap.cxx,v 1.75 2010/02/17 19:00:29 jchiang Exp $
  */
 
 #include <algorithm>
@@ -147,8 +147,8 @@ SourceMap::SourceMap(Source * src, const CountsMap * dataMap,
          // map geometry without adding padding for psf leakage since
          // this cannot be done in general without redefining the
          // reference pixel and reference direction.
-         naxis1 = dataMap->naxis1()*resamp_factor;
-         naxis2 = dataMap->naxis2()*resamp_factor;
+         naxis1 = static_cast<int>(dataMap->naxis1()*resamp_factor);
+         naxis2 = static_cast<int>(dataMap->naxis2()*resamp_factor);
          // Ensure an even number of pixels in each direction.
          if (naxis1 % 2 == 1) {
             naxis1 += 1;
